fix: repair broken imports and ui prefix after docs rename
- Restore lesson note imports (note/note.svelte, quote-selection)
- Fix DocCommentMark extension path (note-comment-mark)
- Apply ui: prefix on site-header responsive classes
