+++ apps/dashboard/src/lib/features/course/components/lesson/index.ts
@@ -3,7 +3,7 @@ export { default as LanguageSelector } from './language-selector.svelte';
 export { default as LessonPageEditHeader } from './lesson-page-edit-header.svelte';
 export { default as Loader } from './loader.svelte';
 export { default as Comments } from './comments.svelte';
-export { default as Note } from './doc/note.svelte';
+export { default as Note } from './note/note.svelte';
 export { default as Slide } from './slide/slide.svelte';
 export { default as Video } from './video/video.svelte';
 export { default as Document } from './document/document.svelte';
