+++ packages/ui/src/custom/editor/index.ts
@@ -1,4 +1,4 @@
 export { default as Editor } from './editor.svelte';
-export { DocCommentMark } from './extensions/doc-comment';
+export { DocCommentMark } from './extensions/note-comment';
 export { TABLE_OF_CONTENTS_INITIAL_CONTENT } from './extensions/table-of-contents';
 export type { HTMLContent, Content, Editor as TiptapEditor, Extensions } from '@tiptap/core';
